Add "select" lazy operation (#3)
* Add "select" lazy operation

This operation performs the "select" verb by selecting or dropping
columns from our SQL query.
It is going to select columns from a subquery if a limit was already
defined.

* Refactor to simplify query planning

We don't need the operations anymore. It is possible to build the query
directly in the Query struct. After that the adapter will build the SQL
version of the query plan.